Added createNextGen method
Crossover: Crossover works fine so far, the only thing to confirm is,
will generating random gene work when the gene is repeated in the child?
The crossover still needs a tournament selection for parents (or any
useful method to ensure better results)
Print methods: We still need a better print method to visualize the best
chromosome.